% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isawesome.R
\name{isawesome}
\alias{isawesome}
\title{Gives compliments}
\usage{
isawesome(name = "Someone", alternative = FALSE, return_message = TRUE)
}
\arguments{
\item{name}{A character string of the person or object to be complimented}

\item{alternative}{A logical indicating whether alternatives to 'is awesome' should be used}

\item{return_message}{A logical indicating whether a message should be returned.  If FALSE a character string will be returned.}
}
\value{
A message that pays a compliment.
}
\description{
Displays a message that pays compliments.
}
\examples{
isawesome("Randy 'Macho Man' Savage")
isawesome("Hulk Hogan", alternative = TRUE)
}
