% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isawesome.R
\name{isawesome}
\alias{isawesome}
\title{Gives compliments}
\usage{
isawesome(name = "Someone", alternative = FALSE)
}
\arguments{
\item{name}{A character string of the person or object to be complimented}

\item{alternative}{A logical indicating whether alternatives to 'is awesome' should be used}
}
\value{
A message that pays `name` a compliment.
}
\description{
`isawesome()` displays a message that compliments the input.
}
\examples{
isawesome("Randy 'Macho Man' Savage")
}
